vistas con componentes completas
